% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGS.R
\name{cv.EMGS}
\alias{cv.EMGS}
\title{Perform EMGS method with cross validation}
\usage{
cv.EMGS(X, v0s, v1, lambda, a, b, epsilon = 1e-05, K = 5, maxitr = 10000,
  copula, copula_itr, group = NULL, ...)
}
\arguments{
\item{X}{n by p matrix of data}

\item{v1}{parameter v1, standard deviation of the slab distribution}

\item{lambda}{parameter for the diagonal elements}

\item{a}{prior for the sparsity parameter pi}

\item{b}{prior for the sparsity parameter pi}

\item{epsilon}{tolerance for the convergence assessment}

\item{K}{number of cross validation folds.}

\item{maxitr}{maximum number of iterations}

\item{copula}{logical indicator of copula model}

\item{group}{grouping information}

\item{...}{other parameters passed to \code{\link{EMGS}}}

\item{v0}{an increasing sequence of parameter v0, standard deviation of the spike distribution}

\item{copular_itr}{number of Monte Carlo iterations in simulating S}
}
\value{
fitted EMGS objects with largest likelihood and corresponding to 1-SE rule.
}
\description{
This function performs EMGS method with cross validation
}
\examples{

}
\references{
Li, Z. R., & McCormick, T. H. (2017). \emph{An Expectation Conditional Maximization approach for Gaussian graphical models}. arXiv preprint arXiv:1709.06970.
}
