% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGS.R
\name{cv.glasso}
\alias{cv.glasso}
\title{Perform Graphical lasso with cross validation}
\usage{
cv.glasso(X, rholist = NULL, K = 5)
}
\arguments{
\item{X}{n by p matrix of data}

\item{rholist}{list of tunning parameters}

\item{K}{number of cross validation folds.}
}
\value{
fitted objects with largest likelihood and corresponding to 1-SE rule.
}
\description{
This function performs graphical lasso with cross validation
}
\examples{

}
