\name{EMGS-package}
\alias{EMGS-package}
\alias{EMGS}
\docType{package}
\title{
\packageTitle{EMGS}
}
\description{
\packageDescription{EMGS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMGS}
\packageIndices{EMGS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EMGS}

Maintainer: \packageMaintainer{EMGS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
