% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMGS.R
\name{EMGS}
\alias{EMGS}
\title{Perform EMGS method}
\usage{
EMGS(X, v0, v1, lambda, a, b, epsilon = 1e-05, maxitr = 10000,
  verbose = TRUE, savepath = FALSE, copula = FALSE, copula_itr = 20,
  group = NULL, a_tau = 2, b_tau = 1, weights = NULL)
}
\arguments{
\item{X}{n by p matrix of data}

\item{v0}{an increasing sequence of parameter v0, standard deviation of the spike distribution}

\item{v1}{parameter v1, standard deviation of the slab distribution}

\item{lambda}{parameter for the diagonal elements}

\item{a}{prior for the sparsity parameter pi}

\item{b}{prior for the sparsity parameter pi}

\item{epsilon}{tolerance for the convergence assessment}

\item{maxitr}{maximum number of iterations}

\item{verbose}{logical indicator of printing information at each iteration}

\item{savepath}{logical indicator of saving the estimator at each iteration in the EM algorithm}

\item{copula}{logical indicator of copula model}

\item{group}{grouping information}

\item{a_tau}{prior for the group scaling parameter}

\item{b_tau}{prior for the group scaling parameter}

\item{weights}{a vector of weights given to each observation}

\item{copular_itr}{number of Monte Carlo iterations in simulating S}
}
\value{
an fitted EMGS object
}
\description{
This function performs EMGS method
}
\examples{

}
\references{
Li, Z. R., & McCormick, T. H. (2017). \emph{An Expectation Conditional Maximization approach for Gaussian graphical models}. arXiv preprint arXiv:1709.06970.
}
