% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{get.ROC}
\alias{get.ROC}
\title{Calculate the true positive rate and false positive rate}
\usage{
get.ROC(obj, G, thres = 0.5)
}
\arguments{
\item{obj}{fitted EMGS object, or a three-dimensional array of precision matrices}

\item{G}{the true graph. Diagonal entries can take any value.}

\item{thres}{threshold of slab}
}
\value{
a matrix of TPR and FPR
}
\description{
This function calculates the true positive rate and false positive rate of edge discovery
}
\examples{

}
