% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{getNorms}
\alias{getNorms}
\title{Calculate the matrix norms}
\usage{
getNorms(obj, prec, cov = NULL)
}
\arguments{
\item{obj}{a fitted EMGS object or a list of precision matrices}

\item{prec}{true precision matrix}

\item{cov}{true covariance matrix}
}
\value{
norms of the bias for (1) precision matrix, (2) standardized precision matrix, (3) covariance matrix, and (4) correlation matrix
}
\description{
This function calculates the M/S/F/Inf norms for the bias of a list of estimated precision matrices with the truth.
}
\examples{

}
