% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{path.plot}
\alias{path.plot}
\title{Plot the regularization path}
\usage{
path.plot(v0s, obj, G, thres = 0.5, normalize = FALSE, xlab = "",
  ylab = "", main = "")
}
\arguments{
\item{v0s}{the vector of parameter v0}

\item{obj}{fitted EMGS object, or a three-dimensional array of precision matrices}

\item{thres}{threshold of slab}

\item{normalize}{logical indicator to plot precision matrices or partial correlation matrices}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{main}{plot title}
}
\value{
the regularization path
}
\description{
This function plots the regularization path
}
\examples{

}
